EVEREST-244-scheduled-backups-list: fix tests
